AXLink hints are now placed near the bottom left corner of the AXLink frame instead of in the middle. Placing the hint in the middle leads to unclickable links when the link wraps only once (see #373).
The hint text size acts as the offset from the corner since it's likely the viewers choice of hint text size correlates to their text viewing size. This means that most of the time the hint will be placed at the beginning of the last line wrap of the link.